feat(doctypes): Add AdministrativeProcedure model
