+++ packages/cozy-doctypes/src/contacts/Contact.js
@@ -125,44 +125,6 @@ class Contact extends Document {
   static getDisplayName(contact) {
     return Contact.getFullname(contact) || Contact.getPrimaryEmail(contact)
   }
-
-  /**
-   * Returns personal data for the contact
-   *
-   * @param {Object} contact - A contact
-   * @param {Array} fields - The list of fields to retrieve
-   * @return {Object} - the personal data
-   **/
-  static getPersonalData(contact, fields) {
-    const mapping = {
-      firstname: {
-        path: 'name.givenName'
-      },
-      lastname: {
-        path: 'name.familyName'
-      },
-      email: {
-        getter: Contact.getPrimaryEmail
-      },
-      phone: {
-        getter: Contact.getPrimaryPhone
-      }
-    }
-    let personalData = {}
-    fields.forEach(field => {
-      const contactField = get(mapping, field, field)
-      let value
-      if (contactField.getter) {
-        value = contactField.getter(contact)
-      } else {
-        value = get(contact, contactField.path, '')
-      }
-
-      personalData[field] = value
-    })
-
-    return personalData
-  }
 }
 
 const ContactShape = PropTypes.shape({
